Cleaner URLs
By giving each directory its own index.html, this allows for cleaner URLs internet facing.
Additionally, in the spirit of the directory, Learning has been updated with notes explaining the code in the file. May create a learning.css to provide the same function as base.css but with the learning notes of the Learning directory.